Move crystal material definition into a separate file for Explorer
