+++ src/ExplorerDetector.cpp
@@ -2,6 +2,7 @@
 #include "ExplorerParameterisationCrystals.h"
 #include "ExplorerParameterisationBlocks.h"
 #include "ExplorerParameterisationPanels.h"
+#include "CrystalMaterial.h"
 
 #include "G4Material.hh"
 #include "G4NistManager.hh"
@@ -22,50 +23,7 @@ G4VPhysicalVolume* ExplorerDetector::Construct( std::string Name, G4LogicalVolum
   // Materials
   G4NistManager* nistManager = G4NistManager::Instance();
   G4Material* air = nistManager->FindOrBuildMaterial( "G4_AIR" );
-  G4bool isotopes = false;
-
-  G4Material* crystal = nullptr;
-
-  G4Element* O  = nistManager->FindOrBuildElement( "O" , isotopes );
-  G4Element* Si = nistManager->FindOrBuildElement( "Si", isotopes );
-  G4Element* Lu = nistManager->FindOrBuildElement( "Lu", isotopes );
-  G4Material* LSO = new G4Material( "Lu2SiO5", 7.4*g/cm3, 3 );
-  LSO->AddElement( Lu, 2 );
-  LSO->AddElement( Si, 1 );
-  LSO->AddElement( O , 5 );
-
-  G4Element* Y  = nistManager->FindOrBuildElement( "Y" , isotopes );
-  G4Material* LYSO = new G4Material( "LYSO", 7.1*g/cm3, 4 );
-  LYSO->AddElement( Lu, 71.447 * perCent );
-  LYSO->AddElement( Y,  4.034  * perCent );
-  LYSO->AddElement( Si, 6.371  * perCent );
-  LYSO->AddElement( O,  18.148 * perCent );
-
-  //Hanna: material available via nistManager, no need to define
-  G4Material* NaI = nistManager->FindOrBuildMaterial("G4_SODIUM_IODIDE");
-
-  //Hanna: material available via nistManager, no need to define
-  G4Material* BGO = nistManager->FindOrBuildMaterial("G4_BGO");
-
-  //Hanna: material available via nistManager, no need to define
-  G4Material* CsF = nistManager->FindOrBuildMaterial("G4_CESIUM_FLUORIDE");
-
-  std::map<std::string, G4Material*> materialMap = {
-    {"LSO", LSO},
-    {"LYSO", LYSO},
-    {"NaI", NaI},
-    {"BGO", BGO}, 
-    {"CsF", CsF}
-  };
-
-  if (auto search = materialMap.find(Material); search != materialMap.end())
-        crystal = search->second;
-  else if (Material.empty())
-    crystal = LYSO;
-  else {
-    std::cerr << "Unrecognised detector material: " << Material << std::endl;
-    exit(1);
-  }
+  G4Material* crystal = CrystalMaterial::GetCrystalMaterial(Material);
 
   std::cout << "Selected detector material: " << crystal->GetName() << std::endl;
